restructure project for packaging
