+++ README.md
@@ -1,39 +1,63 @@
-# securityheaders
+# secheaders
 Python script to check HTTP security headers
+
 
 Same functionality as securityheaders.io but as Python script. Also checks some server/version headers. Written and tested using Python 3.4.
 
 With minor modifications could be used as a library for other projects.
 
+## Installation
+
+The following assumes you have Python  installed and command `python` refers to python version >= 3.4.
+
+### Run without installation
+
+1. Clone into repository
+2. Run `python -m secheaders`
+
+### Installation
+
+1. Clone into repository
+2. `python -m build`
+3. `pip install dist/secheaders-0.1.0-py3-none-any.whl`
+4. Run `secheaders --help`
+
+
+
 ### Usage
 ```
-$ python securityheaders.py --help
-usage: securityheaders.py [-h] [--max-redirects N] URL
+$ secheaders --help
+usage: secheaders [-h] [--max-redirects N] [--no-check-certificate] URL
 
 Check HTTP security headers
 
 positional arguments:
-  URL                Target URL
+  URL                   Target URL
 
-optional arguments:
-  -h, --help         show this help message and exit
-  --max-redirects N  Max redirects, set 0 to disable (default: 2)
-$
+options:
+  -h, --help            show this help message and exit
+  --max-redirects N     Max redirects, set 0 to disable (default: 2)
+  --no-check-certificate
+                        Do not verify TLS certificate chain (default: False)
 ```
 
-### Output
+
+### Example output
 ```
-$ python securityheaders.py --max-redirects 5 https://secfault.fi
-Header 'x-xss-protection' is missing ... [ WARN ]
-Header 'x-content-type-options' is missing ... [ WARN ]
+$ secheaders example.com
+Header 'x-frame-options' is missing ... [ WARN ]
+Header 'strict-transport-security' is missing ... [ WARN ]
 Header 'content-security-policy' is missing ... [ WARN ]
-Header 'x-powered-by' is missing ... [ OK ]
-Header 'x-frame-options' contains value 'DENY' ... [ OK ]
-Header 'strict-transport-security' contains value 'max-age=63072000' ... [ OK ]
-Header 'access-control-allow-origin' is missing ... [ OK ]
-Header 'server' contains value 'nginx/1.10.1' ... [ WARN ]
+Header 'x-content-type-options' is missing ... [ WARN ]
+Header 'x-xss-protection' is missing ... [ OK ]
+Header 'referrer-policy' is missing ... [ WARN ]
+Header 'permissions-policy' is missing ... [ WARN ]
+Header 'server' contains value 'ECAcc (nyd/D187) ... [ WARN ]
 HTTPS supported ... [ OK ]
 HTTPS valid certificate ... [ OK ]
-HTTP -> HTTPS redirect ... [ OK ]
-$
+HTTP -> HTTPS redirect ... [ WARN ]
 ```
+
+## Note
+
+The project renamed (2024-10-19) from **securityheaders** to **secheaders** to avoid confusion with PyPI package with similar name.
