+++ secheaders/__main__.py
@@ -1,0 +1,3 @@
+if __name__ == "__main__":
+    from .securityheaders import main
+    main()
