+++ secheaders/exceptions.py
@@ -1,0 +1,10 @@
+class SecurityHeadersException(Exception):
+    pass
+
+
+class InvalidTargetURL(SecurityHeadersException):
+    pass
+
+
+class UnableToConnect(SecurityHeadersException):
+    pass
